Add hierarchy test and run in CI as with UCO CI
This patch applies the framework UCO uses for its CI to run a `pytest`
and to validate the sample instance data.  UCO's virtual environment and
monolithic ontology file are built by running part of its CI.

The test demonstrates how to search for the sample's phone using either
direct or indirect taxonomy knowledge (i.e. searching for the phone's
exact taxon, or searching for all Android phones, or for all mobile
phones).

References:
* [UCO OC-140] (CP-99) UCO should provide a SKOS taxonomy of device
  types for observable:deviceType

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>