Add custom device sample
This patch adds a sample that includes a device documented through
reverse engineering.  The described device appeared in the academic
literature, and the source article is cited inline in the JSON-LD.

This taxonomy extension demonstrates using a non-UCO taxon, and confirms
it behaves as expected with a sufficiently broad discovery query.  (Note
the last test now asserts a subset relationship between the expected
device types and the found device types.)

References:
* [UCO OC-140] (CP-99) UCO should provide a SKOS taxonomy of device
  types for observable:deviceType

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>